Feature: makes game to run on click
This change makes the game to run on the click of any of the
buttons shown to the player, hence previous "playGame" function
is longer necessary.
It also changes the name of the function "game" to "playRound".
The reason behind this is that now the name of the function is
in line with the name suggested in the odin project.